Add files via upload
A semester's assignment in learning fluency in DX11- please note, starter files were provided that is not my own code (such as PBR.hlsli, DXCore.cpp, and DXStarter files). The textures and models were also given via the course, and are not of my own making. However, a lot of the classes, shaders and other code are of my own making, and it illustrates my ability to work with C++ and DX11.